Pull operator mixins into their own files
