+++ src/pipeline/messages.py
@@ -1,47 +1,83 @@
-"""Helper functions for message processing in pipelines."""
+"""Mixin for topic message processing operators."""
 
 import duckdb
 
 from settings import settings
+from src.di import module
+from src.di.types.base_module import BaseModule
+from src.di.types.data_source import resolve
 from src.message.base import MessageDataset
 from src.pipeline import base
 from src.source.base import SourceFactory
 from src.topic.base import TopicRegistry
 
 
-def to_duckdb(  # noqa: PLR0913
-    factory: SourceFactory,
-    registry: TopicRegistry,
-    dataset: MessageDataset,
-    topics: list[str] | None,
-    asof_seconds: float,
-    lookback: base.Lookback | None = None,
-    ffill: bool = False,
-) -> duckdb.DuckDBPyRelation:
-    """Return a DuckDB relation of topic messages up to `asof_seconds`.
+class TopicMessageMixin:
+    """Mixin for operators that work on messages in a topic."""
 
-    Thin wrapper around `MessageDataset.to_duckdb` to support lookback.
+    _factory: SourceFactory
+    _registry: TopicRegistry
+    _dataset: MessageDataset
 
-    """
-    start_seconds = None
-    if lookback and lookback.unit != base.Unit.FRAME and asof_seconds - lookback.to_seconds() >= 0:
-        start_seconds = asof_seconds - lookback.to_seconds()
+    @property
+    def factory(self) -> SourceFactory:  # noqa: D102
+        return self._factory
 
-    relation = dataset.to_duckdb(
-        factory=factory,
-        registry=registry,
-        topics=topics,
-        start_seconds=start_seconds,
-        end_seconds=asof_seconds,
-        ffill=ffill,
-        empty=False,
-    )
+    @property
+    def registry(self) -> TopicRegistry:  # noqa: D102
+        return self._registry
 
-    if lookback and lookback.unit == base.Unit.FRAME:
-        relation = (
-            relation.order(f"{settings.TIMESTAMP_SECONDS_COLUMN_NAME} DESC")
-            .limit(lookback.last)
-            .order(f"{settings.TIMESTAMP_SECONDS_COLUMN_NAME} ASC")
+    @property
+    def dataset(self) -> MessageDataset:  # noqa: D102
+        return self._dataset
+
+    def setup(self, path: str, **kwargs) -> None:  # noqa: ANN003
+        """Implement `base.Operator.setup`."""
+        ds_type = resolve(path)
+
+        factory_module = f"{BaseModule.SOURCE_FACTORY.value}.{ds_type.value}"
+        registry_module = f"{BaseModule.TOPIC_REGISTRY.value}.{ds_type.value}"
+        dataset_module = f"{BaseModule.MESSAGE_DATASET.value}.{ds_type.value}"
+
+        self._factory = module.provide(factory_module, {"path": path, **kwargs})
+        self._registry = module.provide(registry_module, {**kwargs})
+        self._dataset = module.provide(dataset_module, {**kwargs})
+
+    def to_duckdb(
+        self,
+        topics: list[str] | None,
+        asof_seconds: float,
+        lookback: base.Lookback | None = None,
+        ffill: bool = False,
+    ) -> duckdb.DuckDBPyRelation:
+        """Return a DuckDB relation of topic messages up to `asof_seconds`.
+
+        Thin wrapper around `MessageDataset.to_duckdb` to support lookback.
+
+        """
+        start_seconds = None
+        if (
+            lookback
+            and lookback.unit != base.Unit.FRAME
+            and asof_seconds - lookback.to_seconds() >= 0
+        ):
+            start_seconds = asof_seconds - lookback.to_seconds()
+
+        relation = self.dataset.to_duckdb(
+            factory=self.factory,
+            registry=self.registry,
+            topics=topics,
+            start_seconds=start_seconds,
+            end_seconds=asof_seconds,
+            ffill=ffill,
+            empty=False,
         )
 
-    return relation
+        if lookback and lookback.unit == base.Unit.FRAME:
+            relation = (
+                relation.order(f"{settings.TIMESTAMP_SECONDS_COLUMN_NAME} DESC")
+                .limit(lookback.last)
+                .order(f"{settings.TIMESTAMP_SECONDS_COLUMN_NAME} ASC")
+            )
+
+        return relation
